headline_analyzer/requirements.txt
pandas==2.2.2
numpy==1.26.4
google-cloud-storage==2.17.0
transformers==4.41.2
transformers[torch]==4.41.2
torch==2.3.1
newsapi-python==0.2.7
Python-dotenv==1.0.1
datasets==2.19.2
python-dotenv==1.0.1
nltk==3.8.1